First test version
I still have to understand the assignment goal so i made two different types of producer threads. If i understood correctly, the FileCrawler2 works fine, but the PathReaderThread needs a better condition to keep on waiting for other paths to be added on the ConsistentFilePathBuffer. In fact, i think that multiple PathReaderThreads might bring one of them to wait forever.